Extract paths from gulpfile to paths.js
